+++ packages/browserslist-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atomicsmash/browserslist-config
 
+## 7.0.0
+
+## 6.0.1
+
+## 6.0.0
+
 ## 5.0.0
 
 ## 4.0.0