+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @atomicsmash/eslint-config
 
+## 7.0.0
+
+### Major Changes
+
+- [#104](https://github.com/AtomicSmash/packages/pull/104) [`ea116ea`](https://github.com/AtomicSmash/packages/commit/ea116eafe1ef929db0d4bd2f4981c8976e64d36f) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Remove default dependencies, use peer dependencies as recommended by ESLint
+
+### Patch Changes
+
+- [#104](https://github.com/AtomicSmash/packages/pull/104) [`58ec501`](https://github.com/AtomicSmash/packages/commit/58ec5015665def18ec96ee6c7f3fff2c396ff01d) Thanks [@mikeybinns](https://github.com/mikeybinns)! - prevent package code leaks including node_modules
+
+## 6.0.1
+
+## 6.0.0
+
+### Major Changes
+
+- [#95](https://github.com/AtomicSmash/packages/pull/95) [`1b446f2`](https://github.com/AtomicSmash/packages/commit/1b446f2073511a71f5e9c6eedcf803804db35942) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Update ESlint config with breaking changes
+
 ## 5.0.0
 
 ## 4.0.0